==> 02-genotype-data-exploration.Rmd <==
---
title: "Data visualization and exploration"
author: "Kaitlyn"
date: "2019-04-11"
output: md_document
editor_options:
  chunk_output_type: console
---

Set working directory and load data.

```{r setup, message = F, warning = F}
library(here)
library(tidyverse)
library(scales)

# import cleaned data file
genotypes <- read_csv(here::here("data", "cleaned-triplicates-genotypes-qpcr.csv"))
```

## Some summary statistics

Compare storage, work, and number worked across sample conditions.
```{r compare storage}
# number of working samples for each method
genotypes %>%
  group_by(stor, working) %>%
  summarise(working.count = n())

# number of perfect samples for each method
genotypes %>%
  group_by(stor, perfect) %>%
  summarise(perfect.count = n())

# mean and SD of number working
genotypes %>%
  group_by(stor) %>%
  summarise(mean = mean(n),
            sd = sd(n))
```


In another version, get data from long format to wide format.
```{r long to wide}
# subset only columns consistent across samples (name, condition), then spread
genotypes.spread <- genotypes %>%
  subset(select=c(lab_id, cond, stor, n)) %>% 
  spread(key = stor, value = n)  
  
head(genotypes.spread) # confirm it worked, yay!

nrow(genotypes.spread) # number of unique samples = 112

# summarize the columns
summary(genotypes.spread)

sd(genotypes.spread$cond, na.rm = T)
```

Now count how many are in each condition.
```{r stor method count}
genotypes.stor.count <- genotypes.spread %>%
  group_by(cond) %>%
  summarise(count = n())

genotypes.stor.count
```

#### FIG condition-histogram.pdf
```{r Supplementary Figure condition-histogram}
pdf(here::here("figures", "condition-histogram.pdf"))
ggplot(aes(x = -cond), data = genotypes.spread) + 
  geom_bar() +
  theme_minimal() +
  xlab("Sample Condition") +
  ylab("Count")
dev.off()
```


## Storage method comparison

Take a look at the effect of storage method on genotypes.

A histogram showing the number of samples for that amplified at each number of alleles.
```{r storage and number worked, warning = F, message = F}
# wrapped
ggplot(genotypes, aes(x = n, fill = stor)) +
  geom_histogram() +
  facet_wrap(~ stor) +
  ylab("Number of Samples") +
  xlab("Number of Alleles Amplifying") +
  ggtitle("Breakdown of Markers Working Across Storage Methods")

# side-by-side
ggplot(genotypes, aes(x = n, fill = stor)) +
  geom_histogram(position = "dodge", bins = 11) +
  ylab("Number of samples") +
  xlab("Number of alleles amplifying") +
  theme_bw() +
  theme(panel.grid.minor = element_blank()) +
  scale_x_continuous(breaks = c(0:10))
```

#### FIG storage-number-working.pdf
```{r FIG storage-number-working}
# create labels for storage factor
genotypes$stor <- factor(genotypes$stor, 
                         levels = c("dry", "EtOH", "swab"),
                         labels = c("Dry", "Ethanol", "Swab"))

pdf(here::here("figures", "storage-number-working.pdf"), width = 3, height = 5)
ggplot(genotypes, aes(x = n, fill = stor)) +
  geom_histogram(bins = 11) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.minor = element_blank()) +
  facet_grid(stor ~ .) +
  scale_fill_manual(values = c("#e41a1c", "#377eb8", "#4daf4a")) +
  ylab("Number of samples") +
  xlab("Number of alleles amplifying") +
  scale_x_continuous(breaks = c(0:10))
dev.off()
```



Plot whether sample worked against storage method.
```{r storage method working}
# summarize
working.storage <- genotypes %>%
  group_by(stor) %>%
  count(working)

working.storage

# histogram
ggplot(working.storage, aes(x = working, y = n, fill = stor)) +
  geom_bar(stat = "identity", position = "dodge") +
  ggtitle("Samples Yielding Usable Genotypes (> 8/10 markers)") +
  ylab("Number of Samples")

# another way of viewing it
ggplot(working.storage, aes(x = stor, y = n, fill = working)) +
  geom_bar(stat = "identity") +
  ggtitle("Samples Yielding Usable Genotypes (> 8/10 markers)") +
  ylab("Number of Samples")

# summarize
perfect.storage <- genotypes %>%
  group_by(stor) %>%
  count(perfect)

# histogram
ggplot(perfect.storage, aes(x = perfect, y = n, fill = stor)) +
  geom_bar(stat = "identity", position = "dodge") +
  ggtitle("Samples Yielding Complete Genotypes (10/10 markers)") +
  ylab("Number of Samples")

# another way of viewing it
ggplot(perfect.storage, aes(x = stor, y = n, fill = perfect)) +
  geom_bar(stat = "identity") +
  ggtitle("Samples Yielding Complete Genotypes (10/10 markers)") +
  ylab("Number of Samples")
```

As per Bryan's 9/17/19 suggestion, a look at how individual samples performed with different storage methods.

I never did figure out how to get a legend for the geom_dumbbell plot, so just added manually in Illustrator. This link seems promising, though I didn't investigate fully: https://stackoverflow.com/questions/44653597/adding-a-traditional-legend-to-dumbbell-plot-in-ggaltgeom-dumbbell-in-r

So, for now, just know that red = dry, and blue = swab.
```{r dumbbell plot}
# need package ggalt for this plot
library(ggalt)

# get samples in order by their dry values
genotypes.spread$lab_id <- factor(genotypes.spread$lab_id, levels = genotypes.spread$lab_id[order(genotypes.spread$dry)])

ggplot(genotypes.spread, aes(x = dry, 
                                  xend = swab, 
                                  y = lab_id, 
                                  group = lab_id)) + 
  
  geom_dumbbell(color="gray", # color of bar
                size=.5,
                size_x=2,
                size_xend=2,
                colour_x = "red",
                colour_xend="blue",
                show.legend = T) + # color of end point (swab)
  labs(x=NULL, 
       y=NULL) +
  theme(plot.title = element_text(hjust=0.5, face="bold"),
        plot.background=element_rect(fill="#f7f7f7"),
        panel.background=element_rect(fill="#f7f7f7"),
        panel.grid.minor=element_blank(),
        panel.grid.major.y=element_blank(),
        panel.grid.major.x=element_line(),
        axis.ticks=element_blank(),
        legend.position="top",
        panel.border=element_blank()) +
  xlab("Number of Markers Working") +
  ggtitle("Dry (red) vs Swab (blue)")
```

In this case, green = ethanol.
```{r dumbbell plot 2}
# get samples in order by their EtOH values
genotypes.spread$lab_id <- factor(genotypes.spread$lab_id, levels = genotypes.spread$lab_id[order(genotypes.spread$EtOH)])

ggplot(genotypes.spread, aes(x = EtOH, 
                                  xend = swab, 
                                  y = lab_id, 
                                  group = lab_id)) + 
  
  geom_dumbbell(color="gray", # color of bar
                size=.5,
                size_x=2,
                size_xend=2,
                colour_x = "darkgreen",
                colour_xend="blue",
                show.legend = T) + # color of end point (swab)
  labs(x=NULL, 
       y=NULL) +
  theme(plot.title = element_text(hjust=0.5, face="bold"),
        plot.background=element_rect(fill="#f7f7f7"),
        panel.background=element_rect(fill="#f7f7f7"),
        panel.grid.minor=element_blank(),
        panel.grid.major.y=element_blank(),
        panel.grid.major.x=element_line(),
        axis.ticks=element_blank(),
        legend.position="top",
        panel.border=element_blank()) +
  xlab("Number of Markers Working") +
  ggtitle("Ethanol (green) vs Swab (blue)")
```


## Effect of condition on genotypes

And now a look at the effect of condition.
```{r condition effect, warning = F, message = F}
ggplot(genotypes, aes(x = cond, y = n, col = stor)) +
  geom_jitter() +
  geom_smooth() +
  facet_wrap(~ stor)

ggplot(genotypes, aes(x = perfect, y = cond, fill = stor)) +
  geom_boxplot() +
#  geom_jitter() +
  facet_wrap(~ stor) +
  coord_flip()

# summarize
working.storage.cond <- genotypes %>%
  drop_na(cond) %>%
  group_by(stor, as.factor(cond)) %>%
  count(working)
names(working.storage.cond) <- c("stor", "cond", "working", "n")

# histogram
ggplot(working.storage.cond, aes(x = cond, y = n, fill = working)) +
  geom_bar(stat = "identity") +
  facet_wrap(~ stor) +
  ggtitle("Samples Yielding Usable Genotypes (> 8/10 markers) by Condition & Storage") +
  ylab("Number of Samples") 

# scale to percentage so it's easier to compare across sample types
ggplot(working.storage.cond, aes(x = cond, y = n, fill = working)) +
  geom_bar(stat = "identity", position = "fill") +
  facet_wrap(~ stor) +
#  ggtitle("Samples Yielding Usable Genotypes (> 8/10 markers) by Condition & Storage") +
  ylab("Percent of Samples") +
  xlab("Sample Condition")


# summarize
perfect.storage.cond <- genotypes %>%
  drop_na(cond) %>%
  group_by(stor, as.factor(cond)) %>%
  count(perfect)
names(perfect.storage.cond) <- c("stor", "cond", "Complete", "n")

# histogram
ggplot(perfect.storage.cond, aes(x = cond, y = n, fill = Complete)) +
  geom_bar(stat = "identity") +
  facet_wrap(~ stor) +
  ggtitle("Samples Yielding Complete Genotypes (10/10 markers) by Condition & Storage") +
  ylab("Number of Samples") 

# scale to percentage so it's easier to compare across sample types
ggplot(perfect.storage.cond, aes(x = cond, y = n, fill = Complete)) +
  geom_bar(stat = "identity", position = "fill") +
  facet_wrap(~ stor) +
  ggtitle("Samples Yielding Complete Genotypes (10/10 markers) by Condition & Storage") +
  ylab("Percent of Samples") 
```

## Effect of condition - collapsed categories

Try combining 1 and 1.5, 2 and 2.5, etc to better facilitate comparisons across conditions (since sample size is low for each condition at the fine scale). 
```{r combine condition scores}
genotypes$cond.coarse <- fct_collapse(as.factor(genotypes$cond), 
                                           Slimy = c("1", "1.5"),
                                           Wet = c("2", "2.5"),
                                           Shiny = c("3", "3.5"),
                                           Dull = c("4"))

# try another way where we round UP not down
#genotypes$cond.coarse <- fct_collapse(as.factor(genotypes$cond), 
#                                           Slimy = c("1"),
#                                           Wet = c("2", "1.5"),
#                                           Shiny = c("3", "2.5"),
#                                           Dull = c("4", "3.5"))

# initially had combined NA into 4, but this doesn't seem right, so I dropped them
#genotypes$cond.coarse <- fct_explicit_na(genotypes$cond.coarse, na_level = "Dull")

# summarize
working.storage.cond.coarse <- genotypes %>%
  drop_na(cond.coarse) %>%
  group_by(stor, as.factor(cond.coarse)) %>%
  count(working)
names(working.storage.cond.coarse) <- c("stor", "cond.coarse", "working", "n")

# histogram
ggplot(working.storage.cond.coarse, aes(x = cond.coarse, y = n, fill = working)) +
  geom_bar(stat = "identity") +
  facet_wrap(~ stor) +
  ggtitle("Samples Yielding Usable Genotypes (> 8/10 markers) by Condition & Storage") +
  ylab("Number of Samples") 

# scale to percentage so it's easier to compare across sample types
ggplot(working.storage.cond.coarse, aes(x = cond.coarse, y = n, fill = working)) +
  geom_bar(stat = "identity", position = "fill") +
  facet_wrap(~ stor) +
  ggtitle("Samples Yielding Usable Genotypes (> 8/10 markers) by Condition & Storage") +
  ylab("Percent of Samples") 

# summarize
perfect.storage.cond.coarse <- genotypes %>%
  drop_na(cond.coarse) %>%
  group_by(stor, cond.coarse) %>%
  count(perfect)
names(perfect.storage.cond.coarse) <- c("stor", "cond.coarse", "Complete", "n")

# histogram
ggplot(perfect.storage.cond.coarse, aes(x = cond.coarse, y = n, fill = Complete)) +
  geom_bar(stat = "identity") +
  facet_wrap(~ stor) +
  ggtitle("Samples Yielding Complete Genotypes (10/10 markers) by Condition & Storage") +
  ylab("Number of Samples") 

# scale to percentage so it's easier to compare across sample types
ggplot(perfect.storage.cond.coarse, aes(x = cond.coarse, y = n, fill = Complete)) +
  geom_bar(stat = "identity", position = "fill") +
  facet_wrap(~ stor) +
  ggtitle("Samples Yielding Complete Genotypes (10/10 markers) by Condition & Storage") +
  ylab("Percent of Samples") 

# same thing, but organized differently
ggplot(perfect.storage.cond.coarse, aes(x = stor, y = n, fill = Complete)) +
  geom_bar(stat = "identity", position = "fill") +
  facet_wrap(~ cond.coarse) +
  ggtitle("Samples Yielding Complete Genotypes (10/10 markers) by Condition & Storage") +
  ylab("Percent of Samples")
```

#### FIG condition-genotype-usable.pdf
```{r condition-genotype-complete}
# create labels for completeness
working.storage.cond.coarse$working <- factor(working.storage.cond.coarse$working, 
                         levels = c("no", "yes"),
                         labels = c("Not usable genotype", "Usable genotype"))

# scale to percentage so it's easier to compare across sample types
pdf(here::here("figures", "condition-genotype-usable.pdf"), width = 7, height = 5)
ggplot(working.storage.cond.coarse, aes(x = cond.coarse, y = n, fill = working)) +
  geom_bar(stat = "identity", position = "fill") +
  facet_wrap(~ stor) +
  ylab("Percent of samples") +
  xlab("Sample condition") +
  theme_bw() +
  theme(legend.title = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank())
dev.off()


```

Updated the figure based on reviewer comments.

#### FIG condition-genotype-usable-reviewer.pdf
```{r condition-genotype-complete}
# create labels for completeness
working.storage.cond.coarse$working <- factor(working.storage.cond.coarse$working, 
                         levels = c("no", "yes"),
                         labels = c("Not usable genotype", "Usable genotype"))

# scale to percentage so it's easier to compare across sample types
pdf(here::here("figures", "condition-genotype-usable-v2.pdf"), width = 9, height = 4)
ggplot(working.storage.cond.coarse, aes(x = stor, y = n, fill = working)) +
  geom_bar(stat = "identity", position = "fill") +
  facet_grid(~ cond.coarse) +
  ylab("Proportion of samples yielding usable genotypes") +
  xlab("Storage method") +
  theme_bw() +
  theme(legend.title = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank())
dev.off()

pdf(here::here("figures", "genotype-usable.pdf"), width = 5, height = 4)
ggplot(working.storage.cond.coarse, aes(x = stor, y = n, fill = working)) +
  geom_bar(stat = "identity", position = "fill") +
  ylab("Proportion of samples yielding usable genotypes") +
  xlab("Storage method") +
  theme_bw() +
  theme(legend.title = element_blank(),
        panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank())
dev.off()
```
