==> 03-qpcr-data-exploration.Rmd <==
---
title: "Data visualization and exploration - QPCR data"
author: "Kaitlyn"
date: "2019-04-11"
output: md_document
editor_options:
  chunk_output_type: console
---

Set working directory and load data.

```{r setup, message = F, warning = F}
library(here)
library(tidyverse)
library(rstatix)

# import cleaned data file
genotypes <- read_csv(here::here("data", "cleaned-triplicates-genotypes-qpcr.csv"))
```

## Explore & remove outliers
```{r}
# have a look at values outside of 1.5 * IQR (outliers)
boxplot.stats(genotypes$extracted_conc)$out

# what is the lowest value? 21.64863
min(boxplot.stats(genotypes$extracted_conc)$out)

genotypes$extracted_conc_nooutlier <- NA

# change outliers to NA
for(i in 1:nrow(genotypes)) {
  if(is.na(genotypes$extracted_conc[i])) {
    genotypes$extracted_conc_nooutlier[i] <-  NA
  }
  else if(genotypes$extracted_conc[i] >= min(boxplot.stats(genotypes$extracted_conc)$out)) {
    genotypes$extracted_conc_nooutlier[i] <- NA
  } else {
    genotypes$extracted_conc_nooutlier[i] <- genotypes$extracted_conc[i]
  }
}
```

## PLOTS - qPCR by storage method
```{r concentration by storage}
## BOXPLOTS

# as is
ggplot(genotypes, aes(x = stor, y = extracted_conc, fill = stor)) +
  geom_boxplot() +
  ylab("DNA Concentration") +
  xlab("Storage Method") +
  ggtitle("DNA Concentration Across Storage Methods")

# no outliers
ggplot(genotypes, aes(x = stor, y = extracted_conc_nooutlier, fill = stor)) +
  geom_boxplot() +
  ylab("DNA Concentration") +
  xlab("Storage Method") +
  ggtitle("DNA Concentration Across Storage Methods")

# log scale
ggplot(genotypes, aes(x = stor, y = log(extracted_conc), fill = stor)) +
  geom_boxplot() +
  ylab("DNA Concentration (log)") +
  xlab("Storage Method") +
  ggtitle("DNA Concentration Across Storage Methods")

# histogram
ggplot(genotypes, aes(x = extracted_conc)) +
  geom_histogram() +
  facet_grid(stor ~ .)

# log scale histogram
ggplot(genotypes, aes(x = log(extracted_conc))) +
  geom_histogram() +
  facet_grid(stor ~ .)

# density plot
ggplot(genotypes, aes(x = extracted_conc, col = stor)) +
  geom_density() +
  facet_grid(stor ~ .)

# log density plot
ggplot(genotypes, aes(x = log(extracted_conc), col = stor)) +
  geom_density() +
  facet_grid(stor ~ .)

# no outlier density plot
ggplot(genotypes, aes(x = extracted_conc_nooutlier, col = stor)) +
  geom_density() +
  facet_grid(stor ~ .)

# log no outlier
ggplot(genotypes, aes(x = log(extracted_conc_nooutlier), col = stor)) +
  geom_density() +
  xlab("DNA concentration (log ng/µL)") +
  ylab("Density") +
  theme_classic()
```

#### FIG qpcr-density-storage.pdf
```{r export density diagram}
# create labels for storage factor
genotypes$stor <- factor(genotypes$stor, 
                         levels = c("dry", "EtOH", "swab"),
                         labels = c("Dry", "Ethanol", "Swab"))

pdf(here::here("figures", "qpcr-density-storage.pdf"), width = 5, height = 3)
ggplot(genotypes, aes(x = log(extracted_conc), col = stor)) +
  geom_density() +
  xlab("DNA concentration (log ng/µL)") +
  ylab("Density") +
  theme_classic() +
  theme(legend.title = element_blank()) 

dev.off()


# version with IQR lines - from qpr.summary data frame
pdf(here::here("figures", "qpcr-density-storage-iqr.pdf"), width = 5, height = 3)
ggplot(genotypes, aes(x = log(extracted_conc), col = stor)) +
  geom_density() +
  xlab("DNA concentration (log ng/µL)") +
  ylab("Density") +
  theme_classic() +
  theme(legend.title = element_blank()) +
  geom_vline(xintercept = log(0.779), color = "red") +
  geom_vline(xintercept = log(13.5), color = "red") +
  geom_vline(xintercept = log(0.505), color = "blue") +
  geom_vline(xintercept = log(6.37), color = "blue") +
  geom_vline(xintercept = log(1.29), color = "green") +
  geom_vline(xintercept = log(8.05), color = "green")
dev.off()
```


## Calculate some summary stats
```{r summary stats}
(qpcr.summary <- genotypes %>%
  group_by(stor) %>%
  summarize(mean = mean(extracted_conc, na.rm = TRUE),
            median = median(extracted_conc, na.rm = TRUE),
            sd = sd(extracted_conc, na.rm = TRUE),
            min = min(extracted_conc, na.rm = TRUE),
            max = max(extracted_conc, na.rm = TRUE),
            quantile25 = quantile(extracted_conc, 0.25, na.rm = TRUE),
            quantile75 = quantile(extracted_conc, 0.75, na.rm = TRUE)
            ))

(qpcr.summary.log <- genotypes %>%
  group_by(stor) %>%
  summarize(mean = mean(log(extracted_conc), na.rm = TRUE),
            median = median(log(extracted_conc), na.rm = TRUE),
            sd = sd(log(extracted_conc), na.rm = TRUE),
            min = min(log(extracted_conc), na.rm = TRUE),
            max = max(log(extracted_conc), na.rm = TRUE))
            )

(qpcr.summary.working <- genotypes %>%
  group_by(stor, working) %>%
  summarize(mean = mean(extracted_conc, na.rm = TRUE),
            median = median(extracted_conc, na.rm = TRUE),
            sd = sd(extracted_conc, na.rm = TRUE),
            min = min(extracted_conc, na.rm = TRUE),
            max = max(extracted_conc, na.rm = TRUE)
            ))
```

Quick ANOVA to see if there is a difference in DNA concentration across storage methods (there isn't).

```{r anova comparison}
fit <- aov(log(extracted_conc) ~ stor, data = genotypes)
summary(fit)

# more appropriate - repeated-measures ANOVA

# first clean up
genotypes_clean <- genotypes %>% 
  select(lab_id, stor, extracted_conc) %>% 
  drop_na() %>% 
  mutate(extracted_conc_log = log(extracted_conc))

anova_test(data = genotypes_clean,
           dv = extracted_conc_log,
           wid = lab_id,
           between = stor)
```


## Relationship between genotyping and qPCR

Here, plotting the number of amplifying genotypes vs. DNA concentration (qPCR). It doesn't look like there is much of a relationship, even though we might expect a positive relationship.
```{r genotyping vs qpcr}
ggplot(genotypes, aes(log(extracted_conc), n, col = stor)) +
  geom_jitter()
```

Relationship between concentration and complete/working samples across storage methods
```{r concentration and complete or working}
# working
ggplot(genotypes, aes(x = log(extracted_conc), col = working)) +
  geom_density() +
  facet_grid(stor~.) +
  xlab("Extracted Concentration (log)") +
  ylab("Density")

# complete
ggplot(genotypes, aes(x = log(extracted_conc), col = perfect)) +
  geom_density() +
  facet_grid(stor~.) +
  xlab("Extracted Concentration (log)") +
  ylab("Density")
```

#### FIG qpcr-density-complete-storage.pdf
```{r export density diagram}
# create labels for completeness
genotypes$working <- factor(genotypes$working, 
                         levels = c("no", "yes"),
                         labels = c("Not usable genotype", "Usable genotype"))

pdf(here::here("figures", "qpcr-density-working-storage.pdf"), width = 5, height = 5)
ggplot(genotypes, aes(x = log(extracted_conc), col = working)) +
  geom_density() +
  facet_grid(stor~.) +
  xlab("Extracted concentration (log ng/µL)") +
  ylab("Density") +
  theme_bw() +
  theme(legend.title = element_blank())
dev.off()
```


## qPCR vs condition, by storage method
```{r concentration by storage * condition}
ggplot(genotypes, aes(x = cond, y = log(extracted_conc), col = stor)) +
  geom_jitter() +
  facet_wrap(~ stor) +
  ylab("DNA Concentration (log)") +
  xlab("Sample Condition") 

genotypes$cond.coarse <- fct_collapse(as.factor(genotypes$cond), 
                                           Slimy = c("1", "1.5"),
                                           Wet = c("2", "2.5"),
                                           Shiny = c("3", "3.5"),
                                           Dull = c("4"))

ggplot(genotypes, aes(x = log(extracted_conc), col = working)) +
  geom_density() +
  facet_grid(stor~cond.coarse) +
  xlab("Extracted concentration (log ng/µL)") +
  ylab("Density") +
  theme_bw() +
  theme(legend.title = element_blank())

ggplot(genotypes, aes(x = log(extracted_conc), col = cond.coarse)) +
  geom_density() +
  facet_grid(working~.) +
  xlab("Extracted concentration (log ng/µL)") +
  ylab("Density") +
  theme_bw() +
  theme(legend.title = element_blank())
```

Quick t-tests to see if there is a difference in DNA concentration in samples with usable vs not usable genotypes for each of the three storage methods.

```{r anova comparison}
genotypes.dry.work <- subset(genotypes, stor == "dry" & working == "yes")
genotypes.dry.workno <- subset(genotypes, stor == "dry" & working == "no")
t.test(log(genotypes.dry.work$extracted_conc), log(genotypes.dry.workno$extracted_conc))

genotypes.EtOH.work <- subset(genotypes, stor == "EtOH" & working == "yes")
genotypes.EtOH.workno <- subset(genotypes, stor == "EtOH" & working == "no")
t.test(log(genotypes.EtOH.work$extracted_conc), log(genotypes.EtOH.workno$extracted_conc))

genotypes.swab.work <- subset(genotypes, stor == "swab" & working == "yes")
genotypes.swab.workno <- subset(genotypes, stor == "swab" & working == "no")
t.test(log(genotypes.swab.work$extracted_conc), log(genotypes.swab.workno$extracted_conc))
```
